refactor: Phase 5 [5.7] migrate utils/ — move calc_flops, mps_support, training_progress to utils/, add deprecation shims (BPR)
