+++ src/common/calc_flops/__init__.py
@@ -1,0 +1,1 @@
+"""Shim — canonical location is utils/ (Phase 5)."""
