+++ src/training/runners/category_cv.py
@@ -11,9 +11,9 @@ from models.registry import create_model
 from training.helpers import compute_class_weights, setup_optimizer, setup_scheduler
 from training.runners.category_eval import evaluate
 from training.runners.category_trainer import train
-from common.calc_flops.calculate_model_flops import calculate_model_flops
+from utils.flops import calculate_model_flops
 from tracking import MLHistory, FlopsMetrics, NeuralParams
-from common.mps_support import clear_mps_cache
+from utils.mps import clear_mps_cache
 
 
 def run_cv(
