+++ src/training/runners/next_cv.py
@@ -10,10 +10,10 @@ from models.registry import create_model
 from training.helpers import compute_class_weights, setup_optimizer, setup_scheduler
 from training.runners.next_eval import evaluate
 from training.runners.next_trainer import train
-from common.calc_flops.calculate_model_flops import calculate_model_flops
+from utils.flops import calculate_model_flops
 from tracking import MLHistory, FlopsMetrics, NeuralParams
 from tracking.fold import FoldHistory
-from common.mps_support import clear_mps_cache
+from utils.mps import clear_mps_cache
 
 import torch
 import torch.nn as nn
